requirements.txt
Django==4.0.2
python-decouple==3.6
psycopg2==2.8.6
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0